invoke dynamic method
but patching doesn't work yet

+++ src/ctxt.rs
@@ -188,7 +188,7 @@ pub struct Fct<'ast> {
     pub params_types: Vec<BuiltinType>,
     pub return_type: BuiltinType,
     pub ctor: Option<CtorType>,
-    pub vtable_index: Option<usize>,
+    pub vtable_index: Option<u32>,
     pub initialized: bool,
     pub throws: bool,
